(Declarative Shadow DOM) Use shadowroot attribute value from open tag
Prior to this CL, the "shadowroot" attribute from the <template>
DOM node would be read back upon encountering the closing tag,
to determine whether to attach a shadow root. To better align with
the spec, with this CL, the "shadowroot" token attribute from the
opening <template shadowroot> tag is used instead.

This CL also starts to add testing of declarative shadow DOM.

Bug: 1042130
Change-Id: I0f52a3baa4c1f11d4cce7d1001ec80a31f139e48